CAMEL-23572: camel-tui: Extract DocHelper for download and AsciiDoc conversion
